Extract file_excerpt.rs from turn.rs (parent #680)
Move the file excerpt + content-hash helpers out of turn.rs into a new
src/agent/loop_run/file_excerpt.rs sibling module:

- truncate_on_char_boundary (pub(super))
- utf8_prefix_respecting_cap (pub(super), Issue #636 CB-001)
- open_excerpt_file_nofollow (pub(super), Issue #636 CB-002)
- current_file_hash_for_relative_path (pub(super))
- sha256_hex (pub(super))

No facade re-export (DR3-001). Callers updated:
- scaffold_pipeline.rs: switches `current_file_hash_for_relative_path` /
  `sha256_hex` imports from `super::turn` to `super::file_excerpt`.
- turn.rs: re-imports the 4 production helpers from `super::file_excerpt`
  for in-file callers; `sha256_hex` is gated `#[cfg(test)]` (only test
  fixtures call it directly).

Drops `use sha2::{Digest, Sha256};` from turn.rs (now isolated to the
new module).

turn.rs: 26,852 -> 26,710 (-142 LOC)
Cumulative from 39,489: -12,779 LOC (-32.4%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
