+++ src/agent/loop_run/scaffold_pipeline.rs
@@ -44,6 +44,7 @@ use super::actor_loop_flow::{
 use super::completion_evidence::{RepoEditCategory, classify_repo_edit_path};
 use super::deterministic;
 use super::deterministic_fallback_plan::deterministic_timeout_fallback_plan;
+use super::file_excerpt::{current_file_hash_for_relative_path, sha256_hex};
 use super::interrupt::InterruptFlag;
 use super::lifecycle;
 use super::path_helpers::{normalize_memory_path, sync_package_json_with_existing_lock};
@@ -61,9 +62,9 @@ use super::tool_history::{
     focused_edit_target_already_read, has_successful_non_plan_repo_edit, latest_user_turn_slice,
 };
 use super::turn::{
-    WrittenScaffoldArtifacts, current_file_hash_for_relative_path, extract_filename_with_suffix,
-    last_read_tool_path, latest_turn_preferred_read_edit_target, meaningful_workspace_files,
-    sha256_hex, tool_result_failed, workspace_appears_empty, write_stdout_rendered,
+    WrittenScaffoldArtifacts, extract_filename_with_suffix, last_read_tool_path,
+    latest_turn_preferred_read_edit_target, meaningful_workspace_files, tool_result_failed,
+    workspace_appears_empty, write_stdout_rendered,
 };
 use crate::agent::prompting;
 use crate::agent::recovery;
